+++ src/Umbraco.Web.UI.Client/src/packages/members/member/components/input-member/input-member.context.ts
@@ -1,4 +1,4 @@
-import type { UmbMemberItemModel } from '../../repository/index.js';
+import type { UmbMemberItemModel } from '../../item/types.js';
 import {
 	UMB_MEMBER_PICKER_MODAL,
 	type UmbMemberPickerModalData,
