+++ src/Umbraco.Web.UI.Client/src/packages/members/member/constants.ts
@@ -1,7 +1,9 @@
 export { UMB_MEMBER_ENTITY_TYPE, UMB_MEMBER_ROOT_ENTITY_TYPE } from './entity.js';
 export { UMB_MEMBER_VARIANT_CONTEXT } from './property-dataset-context/member-property-dataset.context-token.js';
+
 export * from './collection/constants.js';
 export * from './entity-actions/constants.js';
+export * from './item/constants.js';
 export * from './repository/constants.js';
 export * from './search/constants.js';
 export * from './workspace/constants.js';
