+++ src/Umbraco.Web.UI.Client/src/packages/members/member/index.ts
@@ -3,6 +3,8 @@ import './components/index.js';
 export * from './collection/index.js';
 export * from './components/index.js';
 export * from './constants.js';
+export * from './item/index.js';
 export * from './paths.js';
 export * from './repository/index.js';
+
 export type * from './types.js';
