+++ src/Umbraco.Web.UI.Client/src/packages/members/member/item/index.ts
@@ -1,0 +1,3 @@
+export { UmbMemberItemRepository } from './repository/index.js';
+
+export type * from './types.js';
