+++ src/Umbraco.Web.UI.Client/src/packages/members/member/item/repository/manifests.ts
