+++ src/Umbraco.Web.UI.Client/src/packages/members/member/item/types.ts
@@ -1,0 +1,1 @@
+export type * from './repository/types.js';
