+++ src/Umbraco.Web.UI.Client/src/packages/members/member/repository/constants.ts
@@ -1,2 +1,1 @@
 export * from './detail/constants.js';
-export * from './item/constants.js';
