+++ src/Umbraco.Web.UI.Client/src/packages/members/member/repository/index.ts
@@ -1,3 +1,2 @@
 export { UmbMemberDetailRepository } from './detail/index.js';
-export { UmbMemberItemRepository, type UmbMemberItemModel } from './item/index.js';
 export * from './validation/index.js';
