+++ src/Umbraco.Web.UI.Client/src/packages/members/member/repository/manifests.ts
@@ -1,5 +1,4 @@
 import { manifests as detailManifests } from './detail/manifests.js';
-import { manifests as itemManifests } from './item/manifests.js';
 import { manifests as validationManifests } from './validation/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [...detailManifests, ...itemManifests, ...validationManifests];
+export const manifests: Array<UmbExtensionManifest> = [...detailManifests, ...validationManifests];
