+++ src/Umbraco.Web.UI.Client/src/packages/members/member/repository/member-repository-base.ts
@@ -1,7 +1,7 @@
 import type { UmbMemberDetailStore } from './detail/member-detail.store.js';
 import { UMB_MEMBER_DETAIL_STORE_CONTEXT } from './detail/member-detail.store.context-token.js';
-import type { UmbMemberItemStore } from './item/member-item.store.js';
-import { UMB_MEMBER_ITEM_STORE_CONTEXT } from './item/member-item.store.context-token.js';
+import type { UmbMemberItemStore } from '../item/repository/member-item.store.js';
+import { UMB_MEMBER_ITEM_STORE_CONTEXT } from '../item/repository/member-item.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbNotificationContext } from '@umbraco-cms/backoffice/notification';
 import { UMB_NOTIFICATION_CONTEXT } from '@umbraco-cms/backoffice/notification';
