+++ src/Umbraco.Web.UI.Client/src/packages/members/member/types.ts
@@ -5,6 +5,8 @@ import type { UmbContentDetailModel, UmbElementValueModel } from '@umbraco-cms/b
 
 export type * from './entity.js';
 export type * from './collection/types.js';
+export type * from './item/types.js';
+
 export interface UmbMemberDetailModel extends UmbContentDetailModel {
 	email: string;
 	entityType: UmbMemberEntityType;
@@ -19,7 +21,7 @@ export interface UmbMemberDetailModel extends UmbContentDetailModel {
 	lastPasswordChangeDate: string | null;
 	memberType: {
 		unique: string;
-		 icon: string;
+		icon: string;
 	};
 	newPassword?: string;
 	oldPassword?: string;
